backend/app/main.py
from fastapi import FastAPI
from app.account.routers import router as account_router
# from app.product.routers.category import router as category_router
# from app.product.routers.product import router as product_router
from app.chat.routers import router as chat_router
from fastapi.middleware.cors import CORSMiddleware
from decouple import config
from fastapi.staticfiles import StaticFiles

app = FastAPI(title="Hi ALL App")

# app.mount("/media", StaticFiles(directory="media"), name="media")
app.add_middleware(
    CORSMiddleware,
    allow_origins=[config("FRONTEND_URL")],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"]
)

@app.get("/")
def root():
    return f"I am root endpoint"


app.include_router(account_router)
app.include_router(chat_router)
# app.include_router(product_router)
