backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.5",
    "email-validator>=2.3.0",
    "fastapi[standard]>=0.128.8",
    "httpx>=0.28.1",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=9.0.2",
    "python-decouple>=3.8",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
    "websockets>=16.0",
]
